handle deprecating branch points (#767)
* do not add skips/replaces entries for deprecated bundles


* remove overwritten channel head before add in overwrite-latest mode - preserve upgrade graph after add


* handle branchpoint deprecation

Remove deprecated bundles from channels of the deprecated bundles,
truncating only when a bundle is not on any channel. Branchpoints (i.e,
branches that have two or more different replacements on different
channels) are removed from any channel that the deprecated bundle was
on, while everything till the first branch point is truncated


* fix unit tests, add function documentation


* remove deprecated channel heads on package delete, simplify populator signature


* optional interface for RemoveOverwrittenChannelHead


* comment changes

+++ pkg/lib/registry/registry_test.go
@@ -24,6 +24,7 @@ import (
 	"github.com/operator-framework/operator-registry/pkg/image"
 	"github.com/operator-framework/operator-registry/pkg/lib/bundle"
 	"github.com/operator-framework/operator-registry/pkg/registry"
+	"github.com/operator-framework/operator-registry/pkg/registry/registryfakes"
 	"github.com/operator-framework/operator-registry/pkg/sqlite"
 	"github.com/operator-framework/operator-registry/pkg/sqlite/sqlitefakes"
 )
@@ -282,9 +283,13 @@ func CreateTestDb(t *testing.T) (*sql.DB, func()) {
 	}
 }
 
-func newUnpackedTestBundle(dir, name string, csvSpec json.RawMessage, annotations registry.Annotations) (string, func(), error) {
+func newUnpackedTestBundle(dir, name string, csvSpec json.RawMessage, annotations registry.Annotations, overwrite bool) (string, func(), error) {
 	bundleDir := filepath.Join(dir, fmt.Sprintf("%s-%s", annotations.PackageName, name))
 	cleanup := func() {
+		os.RemoveAll(bundleDir)
+	}
+
+	if overwrite {
 		os.RemoveAll(bundleDir)
 	}
 	if err := os.Mkdir(bundleDir, 0755); err != nil {
@@ -339,112 +344,16 @@ func newUnpackedTestBundle(dir, name string, csvSpec json.RawMessage, annotation
 }
 
 type bundleDir struct {
+	version     string
 	csvSpec     json.RawMessage
 	annotations registry.Annotations
 }
 
-func TestPackagesFromUnpackedRefs(t *testing.T) {
-	tests := []struct {
-		description string
-		bundles     map[string]bundleDir
-		expected    map[string]registry.Package
-		wantErr     bool
-	}{
-		{
-			description: "InvalidBundle/Empty",
-			bundles: map[string]bundleDir{
-				"bundle-empty": {},
-			},
-			wantErr: true,
-		},
-		{
-			description: "LoadPartialGraph",
-			bundles: map[string]bundleDir{
-				"testoperator-1": {
-					csvSpec: json.RawMessage(`{"version":"1.1.0","replaces":"1.0.0"}`),
-					annotations: registry.Annotations{
-						PackageName:        "testpkg-1",
-						Channels:           "alpha",
-						DefaultChannelName: "stable",
-					},
-				},
-				"testoperator-2": {
-					csvSpec: json.RawMessage(`{"version":"2.1.0"}`),
-					annotations: registry.Annotations{
-						PackageName:        "testpkg-2",
-						Channels:           "stable,alpha",
-						DefaultChannelName: "stable",
-					},
-				},
-			},
-			expected: map[string]registry.Package{
-				"testpkg-1": {
-					Name: "testpkg-1",
-					Channels: map[string]registry.Channel{
-						"alpha": {
-							Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
-								registry.BundleKey{
-									BundlePath: fakeBundlePathFromName("testoperator-1"),
-									Version:    "1.1.0",
-									CsvName:    "testoperator-1",
-								}: nil,
-							},
-						},
-					},
-				},
-				"testpkg-2": {
-					Name: "testpkg-2",
-					Channels: map[string]registry.Channel{
-						"alpha": {
-							Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
-								registry.BundleKey{
-									BundlePath: fakeBundlePathFromName("testoperator-2"),
-									Version:    "2.1.0",
-									CsvName:    "testoperator-2",
-								}: nil,
-							},
-						},
-						"stable": {
-							Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
-								registry.BundleKey{
-									BundlePath: fakeBundlePathFromName("testoperator-2"),
-									Version:    "2.1.0",
-									CsvName:    "testoperator-2",
-								}: nil,
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.description, func(t *testing.T) {
-			tmpdir, err := os.MkdirTemp(".", "tmpdir-*")
-			defer os.RemoveAll(tmpdir)
-			require.NoError(t, err)
-			refs := map[image.Reference]string{}
-			for name, b := range tt.bundles {
-				dir, _, err := newUnpackedTestBundle(tmpdir, name, b.csvSpec, b.annotations)
-				require.NoError(t, err)
-				refs[image.SimpleReference(fakeBundlePathFromName(name))] = dir
-			}
-			pkg, err := packagesFromUnpackedRefs(refs)
-			if tt.wantErr {
-				require.Error(t, err)
-				return
-			}
-			require.NoError(t, err)
-			require.EqualValues(t, tt.expected, pkg)
-		})
-	}
-}
-
 func TestCheckForBundles(t *testing.T) {
 	type step struct {
-		bundles map[string]bundleDir
-		action  int
+		bundles  map[string]bundleDir
+		action   int
+		expected map[string]*registry.Package
 	}
 	const (
 		actionAdd = iota
@@ -471,6 +380,7 @@ func TestCheckForBundles(t *testing.T) {
 								Channels:           "stable,alpha",
 								DefaultChannelName: "stable",
 							},
+							version: "1.1.0",
 						},
 						"unorderedReplaces-1.0.0": {
 							csvSpec: json.RawMessage(`{"version":"1.0.0","replaces":"unorderedReplaces-1.1.0"}`),
@@ -479,6 +389,7 @@ func TestCheckForBundles(t *testing.T) {
 								Channels:           "stable,alpha",
 								DefaultChannelName: "stable",
 							},
+							version: "1.0.0",
 						},
 						"unorderedReplaces-1.2.0": {
 							csvSpec: json.RawMessage(`{"version":"1.2.0","replaces":"unorderedReplaces-1.0.0"}`),
@@ -487,9 +398,50 @@ func TestCheckForBundles(t *testing.T) {
 								Channels:           "alpha",
 								DefaultChannelName: "stable",
 							},
+							version: "1.2.0",
 						},
 					},
 					action: actionAdd,
+					expected: map[string]*registry.Package{
+						"testpkg": {
+							Name: "testpkg",
+							Channels: map[string]registry.Channel{
+								"alpha": {
+									Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
+										registry.BundleKey{
+											BundlePath: "unorderedReplaces-1.0.0",
+											Version:    "1.0.0",
+											CsvName:    "unorderedReplaces-1.0.0",
+										}: nil,
+										registry.BundleKey{
+											BundlePath: "unorderedReplaces-1.1.0",
+											Version:    "1.1.0",
+											CsvName:    "unorderedReplaces-1.1.0",
+										}: nil,
+										registry.BundleKey{
+											BundlePath: "unorderedReplaces-1.2.0",
+											Version:    "1.2.0",
+											CsvName:    "unorderedReplaces-1.2.0",
+										}: nil,
+									},
+								},
+								"stable": {
+									Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
+										registry.BundleKey{
+											BundlePath: "unorderedReplaces-1.0.0",
+											Version:    "1.0.0",
+											CsvName:    "unorderedReplaces-1.0.0",
+										}: nil,
+										registry.BundleKey{
+											BundlePath: "unorderedReplaces-1.1.0",
+											Version:    "1.1.0",
+											CsvName:    "unorderedReplaces-1.1.0",
+										}: nil,
+									},
+								},
+							},
+						},
+					},
 				},
 			},
 			wantErr: fmt.Errorf("added bundle unorderedReplaces-1.0.0 pruned from package testpkg, channel stable: this may be due to incorrect channel head (unorderedReplaces-1.1.0)"),
@@ -505,6 +457,7 @@ func TestCheckForBundles(t *testing.T) {
 								PackageName: "testpkg",
 								Channels:    "stable",
 							},
+							version: "1.0.0",
 						},
 						"ignoreDeprecated-1.1.0": {
 							csvSpec: json.RawMessage(`{"version":"1.1.0","replaces":"ignoreDeprecated-1.0.0"}`),
@@ -512,6 +465,7 @@ func TestCheckForBundles(t *testing.T) {
 								PackageName: "testpkg",
 								Channels:    "stable",
 							},
+							version: "1.1.0",
 						},
 						"ignoreDeprecated-1.2.0": {
 							csvSpec: json.RawMessage(`{"version":"1.2.0","replaces":"ignoreDeprecated-1.1.0"}`),
@@ -519,34 +473,102 @@ func TestCheckForBundles(t *testing.T) {
 								PackageName: "testpkg",
 								Channels:    "stable",
 							},
+							version: "1.2.0",
 						},
 					},
 					action: actionAdd,
+					expected: map[string]*registry.Package{
+						"testpkg": {
+							Name: "testpkg",
+							Channels: map[string]registry.Channel{
+								"stable": {
+									Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
+										registry.BundleKey{
+											BundlePath: "ignoreDeprecated-1.0.0",
+											Version:    "1.0.0",
+											CsvName:    "ignoreDeprecated-1.0.0",
+										}: nil,
+										registry.BundleKey{
+											BundlePath: "ignoreDeprecated-1.1.0",
+											Version:    "1.1.0",
+											CsvName:    "ignoreDeprecated-1.1.0",
+										}: nil,
+										registry.BundleKey{
+											BundlePath: "ignoreDeprecated-1.2.0",
+											Version:    "1.2.0",
+											CsvName:    "ignoreDeprecated-1.2.0",
+										}: nil,
+									},
+								},
+							},
+						},
+					},
 				},
 				{
 					bundles: map[string]bundleDir{
 						"ignoreDeprecated-1.1.0": {},
 					},
 					action: actionDeprecate,
-				},
-				{
-					bundles: map[string]bundleDir{
-						"ignoreDeprecated-1.0.0": {
-							csvSpec: json.RawMessage(`{"version":"1.0.0"}`),
-							annotations: registry.Annotations{
-								PackageName: "testpkg",
-								Channels:    "stable",
-							},
-						},
-						"ignoreDeprecated-1.1.0": {
-							csvSpec: json.RawMessage(`{"version":"1.1.0","replaces":"ignoreDeprecated-1.0.0"}`),
-							annotations: registry.Annotations{
-								PackageName: "testpkg",
-								Channels:    "stable",
+					expected: map[string]*registry.Package{
+						"testpkg": {
+							Name: "testpkg",
+							Channels: map[string]registry.Channel{
+								"stable": {
+									Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
+										registry.BundleKey{
+											BundlePath: "ignoreDeprecated-1.1.0",
+											Version:    "1.1.0",
+											CsvName:    "ignoreDeprecated-1.1.0",
+										}: nil,
+										registry.BundleKey{
+											BundlePath: "ignoreDeprecated-1.2.0",
+											Version:    "1.2.0",
+											CsvName:    "ignoreDeprecated-1.2.0",
+										}: nil,
+									},
+								},
 							},
 						},
 					},
+				},
+				{
+					bundles: map[string]bundleDir{
+						"ignoreDeprecated-1.2.0": {
+							csvSpec: json.RawMessage(`{"version":"1.2.0","replaces":""}`),
+							annotations: registry.Annotations{
+								PackageName:        "testpkg",
+								Channels:           "alpha",
+								DefaultChannelName: "alpha",
+							},
+							version: "1.2.0",
+						},
+					},
 					action: actionOverwrite,
+					expected: map[string]*registry.Package{
+						"testpkg": {
+							Name: "testpkg",
+							Channels: map[string]registry.Channel{
+								"stable": {
+									Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
+										registry.BundleKey{
+											BundlePath: "ignoreDeprecated-1.1.0",
+											Version:    "1.1.0",
+											CsvName:    "ignoreDeprecated-1.1.0",
+										}: nil,
+									},
+								},
+								"alpha": {
+									Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{
+										registry.BundleKey{
+											BundlePath: "ignoreDeprecated-1.2.0-overwrite",
+											Version:    "1.2.0",
+											CsvName:    "ignoreDeprecated-1.2.0",
+										}: nil,
+									},
+								},
+							},
+						},
+					},
 				},
 			},
 		},
@@ -571,59 +593,40 @@ func TestCheckForBundles(t *testing.T) {
 					for deprecate := range step.bundles {
 						require.NoError(t, load.DeprecateBundle(deprecate))
 					}
-				case actionAdd:
+				case actionAdd, actionOverwrite:
+					overwriteRefs := map[string][]string{}
 					refs := map[image.Reference]string{}
 					for name, b := range step.bundles {
-						dir, _, err := newUnpackedTestBundle(tmpdir, name, b.csvSpec, b.annotations)
+						dir, _, err := newUnpackedTestBundle(tmpdir, name, b.csvSpec, b.annotations, true)
 						require.NoError(t, err)
-						refs[image.SimpleReference(name)] = dir
+
+						// refs to be added
+						bundleImage := name
+
+						// bundles to remove for overwrite. Only one per package is permitted.
+						if step.action == actionOverwrite {
+							bundleImage += "-overwrite"
+							img, err := registry.NewImageInput(image.SimpleReference(bundleImage), dir)
+							require.NoError(t, err)
+							overwriteRefs[img.Bundle.Package] = append(overwriteRefs[img.Bundle.Package], name)
+						}
+						refs[image.SimpleReference(bundleImage)] = dir
+
 					}
 					require.NoError(t, registry.NewDirectoryPopulator(
 						load,
 						graphLoader,
 						query,
 						refs,
-						nil,
-						false).Populate(registry.ReplacesMode))
+						overwriteRefs).Populate(registry.ReplacesMode))
 
-					err = checkForBundles(context.TODO(), query, graphLoader, refs)
-					if tt.wantErr == nil {
-						require.NoError(t, err)
-						return
-					}
-					require.EqualError(t, err, tt.wantErr.Error())
-
-				case actionOverwrite:
-					overwriteRefs := map[string]map[image.Reference]string{}
-					refs := map[image.Reference]string{}
-					for name, b := range step.bundles {
-						dir, _, err := newUnpackedTestBundle(tmpdir, name, b.csvSpec, b.annotations)
-						require.NoError(t, err)
-						to := image.SimpleReference(name)
-						refs[image.SimpleReference(name)] = dir
-						refs[to] = dir
-						img, err := registry.NewImageInput(to, dir)
-						require.NoError(t, err)
-						if _, ok := overwriteRefs[img.Bundle.Package]; ok {
-							overwriteRefs[img.Bundle.Package] = map[image.Reference]string{}
-						}
-						overwriteRefs[img.Bundle.Package][to] = dir
-					}
-					require.NoError(t, registry.NewDirectoryPopulator(
-						load,
-						graphLoader,
-						query,
-						nil,
-						overwriteRefs,
-						true).Populate(registry.ReplacesMode))
-
-					err = checkForBundles(context.TODO(), query, graphLoader, refs)
-					if tt.wantErr == nil {
-						require.NoError(t, err)
-						return
-					}
-					require.EqualError(t, err, tt.wantErr.Error())
 				}
+				err = checkForBundles(context.TODO(), query, graphLoader, step.expected)
+				if tt.wantErr == nil {
+					require.NoError(t, err, fmt.Sprintf("%d", step.action))
+					continue
+				}
+				require.EqualError(t, err, tt.wantErr.Error())
 			}
 		})
 	}
@@ -673,5 +676,86 @@ func TestDeprecated(t *testing.T) {
 		isDeprecated, err := isDeprecated(context.TODO(), querier, registry.BundleKey{BundlePath: b})
 		require.NoError(t, err)
 		require.Equal(t, deprecated[b], isDeprecated)
+	}
+}
+
+func TestExpectedGraphBundles(t *testing.T) {
+	testBundle, err := registry.NewBundleFromStrings("testBundle", "0.0.1", "testPkg", "default", "default", "")
+	require.NoError(t, err)
+	testBundle.BundleImage = "testImage"
+	testBundleKey := registry.BundleKey{
+		BundlePath: testBundle.BundleImage,
+		Version:    "0.0.1",
+		CsvName:    testBundle.Name,
+	}
+	newTestPackage := func(name string, channelEntries map[string]registry.BundleKey) *registry.Package {
+		channels := map[string]registry.Channel{}
+		for channelName, node := range channelEntries {
+			if _, ok := channels[channelName]; !ok {
+				channels[channelName] = registry.Channel{
+					Nodes: map[registry.BundleKey]map[registry.BundleKey]struct{}{},
+				}
+			}
+			channels[channelName].Nodes[node] = nil
+		}
+		return &registry.Package{
+			Name:     name,
+			Channels: channels,
+		}
+	}
+
+	tests := []struct {
+		description      string
+		graphLoader      registry.GraphLoader
+		bundles          []*registry.Bundle
+		overwrite        bool
+		wantErr          error
+		wantGraphBundles map[string]*registry.Package
+	}{
+		{
+			description: "GraphLoaderError",
+			graphLoader: &registryfakes.FakeGraphLoader{GenerateStub: func(string) (*registry.Package, error) { return nil, fmt.Errorf("graphLoader error") }},
+			bundles:     []*registry.Bundle{testBundle},
+			wantErr:     fmt.Errorf("graphLoader error"),
+		},
+		{
+			description: "NewPackage",
+			graphLoader: &registryfakes.FakeGraphLoader{GenerateStub: func(string) (*registry.Package, error) { return nil, registry.ErrPackageNotInDatabase }},
+			bundles:     []*registry.Bundle{testBundle},
+			wantGraphBundles: map[string]*registry.Package{
+				"testPkg": newTestPackage("testPkg", map[string]registry.BundleKey{"default": testBundleKey}),
+			},
+		},
+		{
+			description: "OverwriteWithoutFlag",
+			graphLoader: &registryfakes.FakeGraphLoader{GenerateStub: func(string) (*registry.Package, error) {
+				return newTestPackage("testPkg", map[string]registry.BundleKey{"alpha": testBundleKey}), nil
+			}},
+			bundles: []*registry.Bundle{testBundle},
+			wantErr: registry.BundleImageAlreadyAddedErr{ErrorString: fmt.Sprintf("Bundle %s already exists", testBundle.BundleImage)},
+		},
+		{
+			description: "OverwriteWithFlag",
+			graphLoader: &registryfakes.FakeGraphLoader{GenerateStub: func(string) (*registry.Package, error) {
+				return newTestPackage("testPkg", map[string]registry.BundleKey{"alpha": testBundleKey}), nil
+			}},
+			bundles:   []*registry.Bundle{testBundle},
+			overwrite: true,
+			wantGraphBundles: map[string]*registry.Package{
+				"testPkg": newTestPackage("testPkg", map[string]registry.BundleKey{"default": testBundleKey}),
+			},
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.description, func(t *testing.T) {
+			graphBundles, err := expectedGraphBundles(tt.bundles, tt.graphLoader, tt.overwrite)
+			if tt.wantErr != nil {
+				require.EqualError(t, err, tt.wantErr.Error())
+				return
+			}
+			require.NoError(t, err)
+
+			require.EqualValues(t, graphBundles, tt.wantGraphBundles)
+		})
 	}
 }
